Adds note to rake task informing user that it is destructive
The title of the task is a little misleading. The assumption could be
made that the task will just update the data via a reindex, when in fact
it drops the current indexes completely and starts over. This can be
very destuctive if you have a large data set which can takes weeks to
reconstruct.